Moved test from pandas/tests/test_frame.py to pandas/io/test_pytables.py
